+++ packages/config/src/discovery/hardcoded/hardcoded.ts
@@ -1,5 +1,5 @@
 /*
-                      ====== IMPORTANT NOTICE ======
+      ====== IMPORTANT NOTICE ======
 
 EDIT THIS FILE ONLY WHEN YOU KNOW WHAT YOU ARE DOING
 
@@ -12,7 +12,11 @@ DO NOT UPDATE THIS FILE ONLY TO FIX THE TESTS
 UNDERSTAND WHAT YOU ARE DOING BEFORE YOU UPDATE THIS FILE
 */
 
-export const TESTS_HARDCODED = {
-  ARBITRUM_SET_VALIDATOR_COUNT: 7,
-  ARBITRUM_SET_SEQUENCER_COUNT: 3,
+import { ARBITRUM_HARDCODED } from './projects/arbitrumHardcoded'
+
+// This is the entry point for hardcoding values in the config
+// only HARDCODED object should be imported in project's config
+
+export const HARDCODED = {
+  ARBITRUM: ARBITRUM_HARDCODED,
 }
